implements authorization

+++ app/Controllers/Products.php
@@ -11,6 +11,8 @@ class Products extends BaseController
     use ResponseTrait;
     protected $productModel;
     protected $validation;
+    protected $authorization;
+    protected $authentication;
 
     public function __construct()
     {
@@ -20,6 +22,8 @@ class Products extends BaseController
             'product_name' => ['label' => 'Product Name', 'rules' => 'required|min_length[3]'],
             'product_price' => ['label' => 'Product Price', 'rules' => 'required|integer']
         ]);
+        $this->authorization = service('authorization');
+        $this->authentication = service('authentication');
     }
 
 	public function index()
@@ -30,62 +34,74 @@ class Products extends BaseController
 
     public function getProducts()
     {
-        $search = $this->request->getVar('search')['value'];
-        $limit = $this->request->getVar('length');
-        $start = $this->request->getVar('start');
-        $order_index = $this->request->getVar('order')[0]['column'];
-        $order_field = $this->request->getVar('columns')[$order_index]['data'];
-        $order_ascdesc = $this->request->getVar('order')[0]['dir'];
+        if (! $this->authorization->hasPermission('read', $this->authentication->user()->id)){
+            return $this->failForbidden("You don't have permissions to view resources.");
+        } else {
+            $search = $this->request->getVar('search')['value'];
+            $limit = $this->request->getVar('length');
+            $start = $this->request->getVar('start');
+            $order_index = $this->request->getVar('order')[0]['column'];
+            $order_field = $this->request->getVar('columns')[$order_index]['data'];
+            $order_ascdesc = $this->request->getVar('order')[0]['dir'];
 
-        $products_query = $this->productModel->orderBy($order_field, $order_ascdesc);
-        if ($search)
-        {
-            $products_query = $products_query->like('product_name', $search);
+            $products_query = $this->productModel->orderBy($order_field, $order_ascdesc);
+            if ($search)
+            {
+                $products_query = $products_query->like('product_name', $search);
+            }
+            $products_data = $products_query->findAll($limit, $start);
+            $products_filter_total = $products_query->countAll();
+            $products_total = $this->productModel->countAll();
+
+            $callback = array(
+                'draw' => $this->request->getVar('draw'),
+                'recordsTotal' => $products_total,
+                'recordsFiltered' => $products_filter_total,
+                'data' => $products_data,
+            );
+
+            header('Content-Type: application/json');
+            echo json_encode($callback);
         }
-        $products_data = $products_query->findAll($limit, $start);
-        $products_filter_total = $products_query->countAll();
-        $products_total = $this->productModel->countAll();
-
-        $callback = array(
-            'draw' => $this->request->getVar('draw'),
-            'recordsTotal' => $products_total,
-            'recordsFiltered' => $products_filter_total,
-            'data' => $products_data,
-        );
-
-        header('Content-Type: application/json');
-        echo json_encode($callback);
     }
 
     public function addProduct()
-    {    
-        $product_data = [
-            'product_name' => $this->request->getVar('product_name'),
-            'product_price' => $this->request->getVar('product_price')
-        ];
-
-        if ($this->validation->run($product_data))
-        {
-            $this->productModel->insert($product_data);
-            return $this->respondCreated($product_data);
+    {
+        if (! $this->authorization->hasPermission('create', $this->authentication->id)){
+            return $this->failForbidden("You don't have permissions to create new resources.");
         } else {
-            return $this->failValidationErrors($this->validation->getErrors());
+            $product_data = [
+                'product_name' => $this->request->getVar('product_name'),
+                'product_price' => $this->request->getVar('product_price')
+            ];
+    
+            if ($this->validation->run($product_data))
+            {
+                $this->productModel->insert($product_data);
+                return $this->respondCreated($product_data);
+            } else {
+                return $this->failValidationErrors($this->validation->getErrors());
+            }
         }
     }
 
     public function editProduct()
     {
-        $product_data = [
-            'product_name' => $this->request->getVar('product_name'),
-            'product_price' => $this->request->getVar('product_price')
-        ];
-
-        if ($this->validation->run($product_data))
-        {
-            $this->productModel->update($this->request->getVar('product_id'), $product_data);
-            return $this->respondCreated($product_data);
+        if (! $this->authorization->hasPermission('edit', $this->authentication->id)){
+            return $this->failForbidden("You don't have permissions to edit resources.");
         } else {
-            return $this->failValidationErrors($this->validation->getErrors());
+            $product_data = [
+                'product_name' => $this->request->getVar('product_name'),
+                'product_price' => $this->request->getVar('product_price')
+            ];
+
+            if ($this->validation->run($product_data))
+            {
+                $this->productModel->update($this->request->getVar('product_id'), $product_data);
+                return $this->respondCreated($product_data);
+            } else {
+                return $this->failValidationErrors($this->validation->getErrors());
+            }
         }
     }
 
@@ -96,9 +112,13 @@ class Products extends BaseController
             return $this->failNotFound('Product ID cannot be null');
         }
 
-        if ($this->productModel->delete($id))
-        {
-            $this->respondDeleted($id);
+        if (! $this->authorization->hasPermission('delete', $this->authentication->id)){
+            return $this->failForbidden("You don't have permissions to delete resources.");
+        } else {
+            if ($this->productModel->delete($id))
+            {
+                $this->respondDeleted($id);
+            }
         }
     }
 }
